呼び出し時に Player class を経由
